+++ src/components/icons/filled/index.ts
@@ -2,10 +2,14 @@ import HomeFilled from './home.vue';
 import DirectionFilled from './direction.vue';
 import DirectionCurveFilled from './direction_curve.vue';
 import SettingsFilled from './settings.vue';
+import DeleteFilled from './delete.vue';
+import AddBoxFilled from './add_box.vue';
 
 export {
     HomeFilled,
     DirectionFilled,
     DirectionCurveFilled,
-    SettingsFilled
+    SettingsFilled,
+    DeleteFilled,
+    AddBoxFilled
 }