+++ src/types/leaflet-extensions.d.ts
@@ -1,0 +1,7 @@
+import * as L from 'leaflet';
+
+declare module 'leaflet' {
+    interface Layer {
+        _leaflet_id: string;
+    }
+}