+++ src/views/trace_route.vue
@@ -2,10 +2,9 @@
 import { VContainer } from 'vuetify/components/VGrid'
 import { VBanner } from 'vuetify/components/VBanner'
 import { VBtn } from 'vuetify/components/VBtn'
-import { polyline } from 'leaflet';
 
 import Editor from '@/components/editor.vue';
-import { inject, ref } from 'vue';
+import { inject, onMounted, ref } from 'vue';
 import type { MapUtil } from '@/utils/map';
 import { useValhallaStore } from '@/stores/valhalla';
 
@@ -21,9 +20,15 @@ const submit = async () => {
 
     const { polylines, alternatePolylines } = valhallaStore.getPolylines(data.trip, data.alternates);
 
-    map?.drawPolyLines(map.map.value, polylines, alternatePolylines);
+    map?.addPolyLines(polylines);
+    alternatePolylines.forEach((alternate) => {
+        map?.addPolyLines(alternate, {
+            color: 'red',
+        });
+    });
 }
 
+onMounted(() => map?.clearAllLayers());
 </script>
 <template>
     <VBanner text="Trace Route"></VBanner>