+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-## 0.1.0 (2024-11-12)
+## 0.2.0 (2024-11-16)
 
 ### Feat
 
+- update application title, enhance layer management, and add isochrone API support
+- add location component, locate API integration, and isochrone screen added
+- add new views and icons, update routing, and enhance map functionality
 - add Dockerfile for multi-stage build and update import path for App component
 - update GitHub Actions workflow to trigger on tag pushes
 - add GitHub Actions workflow for Docker image build and deployment; refactor Vue components to remove unused template references